Add files via upload
file test.py is done by qingtong guo and tianhao zhang, we do the preprocess over the raw data, we divide the data set into two parts: 30% for the test data, 70% for the training data. Fix the not existing data in the annotation.txt, separate the label and fixes of each images, and transfer the fixes of images and labels to the cnn model built by shiqiang.